+++ RELEASE_NOTES.md
@@ -7,6 +7,6 @@ This is the offical release of vpinfe v2.0.7
 None
 
 ### Fixes
-- fix contention bug when the condition is a new install. 
+-  remove Edge browser choice and fat builds use the embeded chromium only!
 
 ### Notes